Formula 1 Logistic Regression Analysis
This model is a classification analysis to predict classification of a Formula 1 driver as "champion" or "not champion". The analysis also uses the LogisticRegression class model.predict_proba method to determine the probabilities associated with the classification of "champion" or "not champion". This is a train only model as I did not have a test set of data.  The probability predictions in my opinion are very useful and this model can be utilized to analyze any sport.   Feel free to leave comments or contact me to discuss.